Add special blocks for tour operator features
- Introduced new block variations for special features including duration, gallery, price type, price, tagline, and booking validity.
- Each block variation is scoped to the 'special' post type and includes metadata, attributes, and inner blocks for structured content.
- Removed deprecated blocks for booking validity and special price.
- Updated asset files for new blocks and ensured proper dependencies are set.
