+++ src/blocks/booking-validity/block.json
@@ -1,0 +1,12 @@
+{
+  "$schema": "https://schemas.wp.org/trunk/block.json",
+  "apiVersion": 3,
+  "textdomain": "to-specials",
+  "name": "lsx-tour-operator/special-booking-validity",
+  "title": "Special - Booking Validity",
+  "category": "lsx-tour-operator",
+  "editorScript": "file:index.js",
+  "description": "Displays the booking validity dates for this special.",
+  "icon": "calendar",
+  "keywords": ["booking", "validity", "dates", "special"]
+}
