+++ src/blocks/duration/block.json
@@ -1,0 +1,12 @@
+{
+  "$schema": "https://schemas.wp.org/trunk/block.json",
+  "apiVersion": 3,
+  "textdomain": "to-specials",
+  "name": "lsx-tour-operator/special-duration",
+  "title": "Special - Duration",
+  "category": "lsx-tour-operator",
+  "editorScript": "file:index.js",
+  "description": "Displays the duration for this special.",
+  "icon": "clock",
+  "keywords": ["duration", "length", "days", "special"]
+}
