--- src/blocks/special-gallery/block.json
@@ -1,12 +1,0 @@
-{
-  "$schema": "https://schemas.wp.org/trunk/block.json",
-  "apiVersion": 3,
-  "textdomain": "to-specials",
-  "name": "lsx-tour-operator/special-gallery",
-  "title": "Special Gallery",
-  "category": "lsx-tour-operator",
-  "editorScript": "file:index.js",
-  "description": "Displays the gallery images for this special.",
-  "icon": "format-gallery",
-  "keywords": ["gallery", "images", "special", "photos"]
-}
